Fix /MIR when using server connection
When purging files over a server connection, the first call to
purgeFilesInDirectory is to the root directory (EACopyClient, line 175).
There, the relPath is "". If there is a connection to the server, the
early return condition in line 1042 kicks in, since that directory "" is
never in m_handledFiles.

Apparently, the second part of the if-clause in 1043 is meant for
purging specific directories via "/PURGE xxx", so it needs to be
protected for the root-directory (aka "").

Repro-case of the bug:

- start EACopyService
- create a folder, e.g. "C:\asd"
- create a shared folder, e.g. "C:\share" and share it to network
- call "EACopy /MIR /SERVER c:\asd \\mypcname\share" (of course,
  assuming your pc is called "mypcname" and the share is called "share")
- observe that nothing is copied into C:\share