/XD will ignore target directories when using /MIR
For vanilla robocopy, the /XD also ignores directories
in the target folder when purging files via /MIR or
/PURGE, so do the same for EACopy.

(This is especially useful to not remove temporary
directories like .git or .vs)

Note, that subdirectories ARE purged if their parent
folder is purged, so if you have foo/bar/baz and specify
/XD baz, but either foo or bar is not part of the source,
then baz will get deleted. This is the same behaviour
as for robocopy.